backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1 //INSTALAR EXPLICITAMENTE
Flask-SQLAlchemy==3.1.1 //INSTALAR EXPLICITAMENTE
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10 //INSTALAR EXPLICITAMENTE
PyJWT==2.10.1 //INSTALAR EXPLICITAMENTE
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
